[trunk] Remove old function tgt_create Fix some warnings about size_t

+++ src/lib/openjp2/event.c
@@ -97,7 +97,7 @@ opj_bool opj_event_msg(opj_common_ptr cinfo, int event_type, const char *fmt, ..
 
 	if ((fmt != NULL) && (event_mgr != NULL)) {
 		va_list arg;
-		int str_length/*, i, j*/; /* UniPG */
+		size_t str_length/*, i, j*/; /* UniPG */
 		char message[MSG_SIZE];
 		memset(message, 0, MSG_SIZE);
 		/* initialize the optional parameter list */
@@ -148,7 +148,7 @@ opj_bool opj_event_msg_v2(opj_event_mgr_t* p_event_mgr, int event_type, const ch
 
 	if ((fmt != 00) && (p_event_mgr != 00)) {
 		va_list arg;
-		int str_length/*, i, j*/; /* UniPG */
+		size_t str_length/*, i, j*/; /* UniPG */
 		char message[MSG_SIZE];
 		memset(message, 0, MSG_SIZE);
 		/* initialize the optional parameter list */
